+++ demo/jest.config.ts
@@ -147,7 +147,7 @@ const config: Config = {
     // setupFiles: [],
 
     // A list of paths to modules that run some code to configure or set up the testing framework before each test
-    setupFilesAfterEnv: ["jest-canvas-mock", "./src/setup-jest.js"],
+    setupFilesAfterEnv: ["jest-canvas-mock", "./src/setup-jest.tsx"],
 
     // The number of seconds after which a test is considered as slow and reported as such in the results.
     // slowTestThreshold: 5,
@@ -165,10 +165,10 @@ const config: Config = {
     // testLocationInResults: false,
 
     // The glob patterns Jest uses to detect test files
-    // testMatch: [
-    //   "**/__tests__/**/*.[jt]s?(x)",
-    //   "**/?(*.)+(spec|test).[tj]s?(x)"
-    // ],
+    testMatch: [
+        "**/__tests__/**/*.(spec|test).[jt]s?(x)",
+        "**/?(*.)+(spec|test).[tj]s?(x)",
+    ],
 
     // An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
     // testPathIgnorePatterns: [