+++ demo/src/__tests__/MemoryModelsUserInput.spec.tsx
@@ -8,6 +8,7 @@ import {
     act,
 } from "@testing-library/react";
 import MemoryModelsUserInput from "../MemoryModelsUserInput.js";
+import { renderWithI18n } from "../setup-jest";
 
 describe("MemoryModelsUserInput", () => {
     // submit button by default resets the form https://stackoverflow.com/a/62404526
@@ -30,15 +31,17 @@ describe("MemoryModelsUserInput", () => {
 
     it("renders Menu for MemoryModelsConfigInput", () => {
         render(
-            <MemoryModelsUserInput
-                onTextDataSubmit={onSubmitMock}
-                setTextData={setTextDataMock}
-                textData={textDataMock}
-                setFailureBanner={setFailureBannerMock}
-                jsonResult={jsonResult}
-                configData={configDataMock}
-                setConfigData={setConfigDataMock}
-            />
+            renderWithI18n(
+                <MemoryModelsUserInput
+                    onTextDataSubmit={onSubmitMock}
+                    setTextData={setTextDataMock}
+                    textData={textDataMock}
+                    setFailureBanner={setFailureBannerMock}
+                    jsonResult={jsonResult}
+                    configData={configDataMock}
+                    setConfigData={setConfigDataMock}
+                />
+            )
         );
         expect(
             screen.getByTestId("rendering-options-menu").textContent
@@ -47,15 +50,17 @@ describe("MemoryModelsUserInput", () => {
 
     it("does not submit the form or enable the submit button with empty textData", () => {
         render(
-            <MemoryModelsUserInput
-                onTextDataSubmit={onSubmitMock}
-                setTextData={setTextDataMock}
-                textData={textDataMock}
-                setFailureBanner={setFailureBannerMock}
-                jsonResult={jsonResult}
-                configData={configDataMock}
-                setConfigData={setConfigDataMock}
-            />
+            renderWithI18n(
+                <MemoryModelsUserInput
+                    onTextDataSubmit={onSubmitMock}
+                    setTextData={setTextDataMock}
+                    textData={textDataMock}
+                    setFailureBanner={setFailureBannerMock}
+                    jsonResult={jsonResult}
+                    configData={configDataMock}
+                    setConfigData={setConfigDataMock}
+                />
+            )
         );
 
         const button = screen.getByTestId("input-submit-button");
@@ -67,15 +72,17 @@ describe("MemoryModelsUserInput", () => {
 
     it("accepts changes to formData", () => {
         render(
-            <MemoryModelsUserInput
-                onTextDataSubmit={onSubmitMock}
-                setTextData={setTextDataMock}
-                textData={textDataMock}
-                setFailureBanner={setFailureBannerMock}
-                jsonResult={jsonResult}
-                configData={configDataMock}
-                setConfigData={setConfigDataMock}
-            />
+            renderWithI18n(
+                <MemoryModelsUserInput
+                    onTextDataSubmit={onSubmitMock}
+                    setTextData={setTextDataMock}
+                    textData={textDataMock}
+                    setFailureBanner={setFailureBannerMock}
+                    jsonResult={jsonResult}
+                    configData={configDataMock}
+                    setConfigData={setConfigDataMock}
+                />
+            )
         );
 
         const updateFormData = "Updated form data";
@@ -89,15 +96,17 @@ describe("MemoryModelsUserInput", () => {
         beforeEach(() => {
             textDataMock = "Form data";
             render(
-                <MemoryModelsUserInput
-                    onTextDataSubmit={onSubmitMock}
-                    setTextData={setTextDataMock}
-                    textData={textDataMock}
-                    setFailureBanner={setFailureBannerMock}
-                    jsonResult={jsonResult}
-                    configData={configDataMock}
-                    setConfigData={setConfigDataMock}
-                />
+                renderWithI18n(
+                    <MemoryModelsUserInput
+                        onTextDataSubmit={onSubmitMock}
+                        setTextData={setTextDataMock}
+                        textData={textDataMock}
+                        setFailureBanner={setFailureBannerMock}
+                        jsonResult={jsonResult}
+                        configData={configDataMock}
+                        setConfigData={setConfigDataMock}
+                    />
+                )
             );
         });
 
@@ -117,15 +126,17 @@ describe("MemoryModelsUserInput", () => {
     describe("MemoryModelsFileInput", () => {
         beforeEach(() => {
             render(
-                <MemoryModelsUserInput
-                    onTextDataSubmit={onSubmitMock}
-                    setTextData={setTextDataMock}
-                    textData={textDataMock}
-                    setFailureBanner={setFailureBannerMock}
-                    jsonResult={jsonResult}
-                    configData={configDataMock}
-                    setConfigData={setConfigDataMock}
-                />
+                renderWithI18n(
+                    <MemoryModelsUserInput
+                        onTextDataSubmit={onSubmitMock}
+                        setTextData={setTextDataMock}
+                        textData={textDataMock}
+                        setFailureBanner={setFailureBannerMock}
+                        jsonResult={jsonResult}
+                        configData={configDataMock}
+                        setConfigData={setConfigDataMock}
+                    />
+                )
             );
         });
 
@@ -256,15 +267,17 @@ describe("MemoryModelsUserInput", () => {
     describe("MemoryModelsConfigInput", () => {
         beforeEach(() => {
             render(
-                <MemoryModelsUserInput
-                    onTextDataSubmit={onSubmitMock}
-                    setTextData={setTextDataMock}
-                    textData={textDataMock}
-                    setFailureBanner={setFailureBannerMock}
-                    jsonResult={jsonResult}
-                    configData={configDataMock}
-                    setConfigData={setConfigDataMock}
-                />
+                renderWithI18n(
+                    <MemoryModelsUserInput
+                        onTextDataSubmit={onSubmitMock}
+                        setTextData={setTextDataMock}
+                        textData={textDataMock}
+                        setFailureBanner={setFailureBannerMock}
+                        jsonResult={jsonResult}
+                        configData={configDataMock}
+                        setConfigData={setConfigDataMock}
+                    />
+                )
             );
         });