+++ webstepper/src/index.tsx
@@ -2,6 +2,7 @@ import React, { StrictMode, useState } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.js";
 import "./css/styles.scss";
+import "./i18n";
 
 import { createTheme, ThemeProvider } from "@mui/material/styles";
 import { useMediaQuery, CssBaseline } from "@mui/material";